{"type":"TI", "runs":"0"}

+++ exerciseCh5.java
@@ -17,7 +17,7 @@ public class exerciseCh5 extends ConsoleProgram {
 		int a = readInt ("a: ");
 		int b = readInt ("b: ");
 		int c = readInt ("c: ");
-		int condition = Math.sqrt((b*b)-(4*a*c));
+		int condition = (int) Math.sqrt((b*b)-(4*a*c));
 		int firstSolution;
 		int secondSolution;
 		if (condition>0) {
